refactor /internal/metrics
