+++ internal/metrics/collector_test.go
@@ -10,7 +10,7 @@ import (
 
 // TestCollector_Collect ensures that Collect() populates memory stats correctly
 func TestCollector_Collect(t *testing.T) {
-	collector := &Collector{}
+	collector := NewCollector()
 	out := make(map[string]any)
 
 	collector.Collect(out)
@@ -27,20 +27,20 @@ func TestCollector_Collect(t *testing.T) {
 	assert.Equal(t, initialPollCount+1, out["PollCount"], "PollCount did not increment")
 }
 
-// TestCollector_RandomValue ensures RandomValue is a float64
+// TestCollector_RandomValue ensures RandomValue is a Gauge type
 func TestCollector_RandomValue(t *testing.T) {
-	collector := &Collector{}
+	collector := NewCollector()
 	out := make(map[string]any)
 
 	collector.Collect(out)
 
 	_, ok := out["RandomValue"].(Gauge)
-	assert.True(t, ok, "RandomValue should be a float64")
+	assert.True(t, ok, "RandomValue should be a Gauge type")
 }
 
 // TestCollector_PollCountIncrements ensures PollCount increments correctly
 func TestCollector_PollCountIncrements(t *testing.T) {
-	collector := &Collector{}
+	collector := NewCollector()
 	out := make(map[string]any)
 
 	collector.Collect(out)
@@ -52,23 +52,9 @@ func TestCollector_PollCountIncrements(t *testing.T) {
 	assert.Equal(t, firstPollCount+1, secondPollCount, "PollCount did not increment properly")
 }
 
-// TestExists_ValidMetrics ensures Exists() returns true for valid metrics
-func TestExists_ValidMetrics(t *testing.T) {
-	for _, metric := range memStats {
-		assert.True(t, Exists(metric), "Exists should return true for valid metric: %s", metric)
-	}
-	assert.True(t, Exists("PollCount"), "PollCount should exist")
-	assert.True(t, Exists("RandomValue"), "RandomValue should exist")
-}
-
-// TestExists_InvalidMetrics ensures Exists() returns false for unknown metrics
-func TestExists_InvalidMetrics(t *testing.T) {
-	assert.False(t, Exists("unknown"), "Exists should return false for unknown metric")
-}
-
 // TestCollector_CorrectDataTypes ensures all collected values are of correct types
 func TestCollector_CorrectDataTypes(t *testing.T) {
-	collector := &Collector{}
+	collector := NewCollector()
 	out := make(map[string]any)
 	collector.Collect(out)
 
@@ -80,7 +66,7 @@ func TestCollector_CorrectDataTypes(t *testing.T) {
 		field := r.FieldByName(stat)
 		switch field.Kind() {
 		case reflect.Uint64, reflect.Uint32, reflect.Float64:
-			assert.IsType(t, Gauge(0), out[stat], "Metric %s has an unexpected type %T", stat, stat)
+			assert.IsType(t, Gauge(0), out[stat], "Metric %s has an unexpected type %T", stat, out[stat])
 		}
 	}
 }
